Refactor and make UI consistent
Share header between overview and practice page.
